+++ previous-versions/sync-for-expenses-version-1/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [typescript v0.5.0] previous-versions/sync-for-expenses-version-1
 ### Releases
 - [NPM v0.5.0] https://www.npmjs.com/package/@codat/sync-for-expenses-version-1/v/0.5.0 - previous-versions/sync-for-expenses-version-1
+
+## 2023-10-02 00:16:01
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses-v1.yaml
+- Speakeasy CLI 1.92.3 (2.143.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.5.1] previous-versions/sync-for-expenses-version-1
+### Releases
+- [NPM v0.5.1] https://www.npmjs.com/package/@codat/sync-for-expenses-version-1/v/0.5.1 - previous-versions/sync-for-expenses-version-1